make cost field a float field
git-svn-id: svn://svn.cy55.de/Zope3/src/cybertools/trunk@3592 fd906abe-77d9-0310-91a1-e0d9ade77398

+++ organize/interfaces.py
@@ -222,7 +222,7 @@ class IService(Interface):
                 title=_(u'External ID'),
                 description=_(u'Identifier in external system or code number.'),
                 required=False,)
-    cost = schema.Int(
+    cost = schema.Float(
                 title=_(u'Cost'),
                 description=_(u'Cost or prizing information.'),
                 required=False,)
